Nodejs changes

+++ sdk/src/types/block/protocol.rs
@@ -55,28 +55,22 @@ pub struct ProtocolParameters {
     #[getset(skip)]
     pub(crate) mana_structure: ManaStructure,
     /// The unbonding period in epochs before an account can stop staking.
-    #[cfg_attr(feature = "serde", serde(with = "crate::utils::serde::string"))]
-    pub(crate) staking_unbonding_period: u64,
+    pub(crate) staking_unbonding_period: u32,
     /// The number of validation blocks that each validator should issue each slot.
     pub(crate) validation_blocks_per_slot: u16,
     /// The number of epochs worth of Mana that a node is punished with for each additional validation block it issues.
-    #[cfg_attr(feature = "serde", serde(with = "crate::utils::serde::string"))]
-    pub(crate) punishment_epochs: u64,
+    pub(crate) punishment_epochs: u32,
     /// Liveness Threshold is used by tip-selection to determine if a block is eligible by evaluating issuingTimes and
     /// commitments in its past-cone to Accepted Tangle Time and lastCommittedSlot respectively.
-    #[cfg_attr(feature = "serde", serde(with = "crate::utils::serde::string"))]
-    pub(crate) liveness_threshold: u64,
+    pub(crate) liveness_threshold: u32,
     /// Minimum age relative to the accepted tangle time slot index that a slot can be committed.
-    #[cfg_attr(feature = "serde", serde(with = "crate::utils::serde::string"))]
-    pub(crate) min_committable_age: u64,
+    pub(crate) min_committable_age: u32,
     /// Maximum age for a slot commitment to be included in a block relative to the slot index of the block issuing
     /// time.
-    #[cfg_attr(feature = "serde", serde(with = "crate::utils::serde::string"))]
-    pub(crate) max_committable_age: u64,
+    pub(crate) max_committable_age: u32,
     /// Epoch Nearing Threshold is used by the epoch orchestrator to detect the slot that should trigger a new
     /// committee selection for the next and upcoming epoch.
-    #[cfg_attr(feature = "serde", serde(with = "crate::utils::serde::string"))]
-    pub(crate) epoch_nearing_threshold: u64,
+    pub(crate) epoch_nearing_threshold: u32,
     /// Parameters used to calculate the Reference Mana Cost (RMC).
     pub(crate) congestion_control_parameters: CongestionControlParameters,
     /// Defines the parameters used to signal a protocol parameters upgrade.
@@ -132,7 +126,7 @@ impl ProtocolParameters {
         token_supply: u64,
         genesis_unix_timestamp: u64,
         slot_duration_in_seconds: u8,
-        epoch_nearing_threshold: u64,
+        epoch_nearing_threshold: u32,
     ) -> Result<Self, Error> {
         Ok(Self {
             version,
